planner, executor: address mlog stats review comments
