+++ pkg/planner/core/casetest/mview/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
         "mv_refresh_test.go",
     ],
     flaky = True,
-    shard_count = 16,
+    shard_count = 18,
     deps = [
         "//pkg/domain",
         "//pkg/infoschema",
